rename dm.py to bmc.py and add new irisao module
