+++ zygo_automation/__init__.py
@@ -1,1 +1,1 @@
-from . import automation, zygo, dm
+from . import automation, zygo, bmc, irisao
