+++ zygo_automation/automation.py
@@ -7,7 +7,7 @@ import h5py
 import numpy as np
 
 from .zygo import capture_frame, read_many_raw_datx
-from .dm import load_channel, set_pixel, set_row_column, write_fits
+from .bmc import load_channel, set_pixel, set_row_column, write_fits
 
 import logging
 logging.basicConfig(level=logging.INFO)
